+++ src/support/utils.js
@@ -47,6 +47,7 @@ export const isAuditForAllDeliveryTypes = (deliveryType) => deliveryType.toUpper
  * @param {string} type - The type of audit.
  * @param {Object} auditContext - The audit context object.
  * @param {string} siteId - The site ID to audit.
+ * @param {string} [auditData] - Optional audit data.
  * @returns {Promise} A promise representing the message sending operation.
  */
 export const sendAuditMessage = async (
@@ -55,7 +56,13 @@ export const sendAuditMessage = async (
   type,
   auditContext,
   siteId,
-) => sqs.sendMessage(queueUrl, { type, siteId, auditContext });
+  auditData,
+) => sqs.sendMessage(queueUrl, {
+  type,
+  siteId,
+  auditContext,
+  data: auditData,
+});
 
 // todo: prototype - untested
 /* c8 ignore start */
@@ -176,6 +183,7 @@ export const sendAuditMessages = async (
  * Triggers an audit for a site.
  * @param {Site} site - The site to audit.
  * @param {string} auditType - The type of audit.
+ * @param {undefined|string} auditData - Optional audit data.
  * @param {Object} slackContext - The Slack context object.
  * @param {Object} lambdaContext - The Lambda context object.
  * @return {Promise} - A promise representing the audit trigger operation.
@@ -183,6 +191,7 @@ export const sendAuditMessages = async (
 export const triggerAuditForSite = async (
   site,
   auditType,
+  auditData,
   slackContext,
   lambdaContext,
 ) => sendAuditMessage(
@@ -196,6 +205,7 @@ export const triggerAuditForSite = async (
     },
   },
   site.getId(),
+  auditData,
 );
 
 // todo: prototype - untested